+++ app/src/main/java/com/zapbot/android/ZapBotApplication.kt
@@ -37,7 +37,8 @@ class AppContainer(app: Application) {
             AppDatabase.MIGRATION_3_4,
             AppDatabase.MIGRATION_4_5,
             AppDatabase.MIGRATION_5_6,
-            AppDatabase.MIGRATION_6_7
+            AppDatabase.MIGRATION_6_7,
+            AppDatabase.MIGRATION_7_8
         )
         .build()
     val moshi: Moshi = Moshi.Builder().add(KotlinJsonAdapterFactory()).build()